move dev code into dev module
