+++ packages/kit/src/vite/utils.js
@@ -157,40 +157,3 @@ export function format_illegal_import_chain(stack) {
 
 	return `Cannot import ${stack.at(-1)} into client-side code:\n${pyramid}`;
 }
-
-/**
- * Throw an error if a private module is imported from a client-side node.
- * @param {import('vite').ModuleNode} node
- * @param {Set<string>} illegal_imports
- */
-export function throw_if_illegal_private_import_vite(node, illegal_imports) {
-	const seen = new Set();
-
-	/**
-	 * @param {import('vite').ModuleNode} node
-	 * @returns {string[] | null}
-	 */
-	function find(node) {
-		if (!node.id) return null; // TODO when does this happen?
-
-		if (seen.has(node.id)) return null;
-		seen.add(node.id);
-
-		if (node.id && illegal_imports.has(node.id)) {
-			return [node.id];
-		}
-
-		for (const child of node.importedModules) {
-			const chain = child && find(child);
-			if (chain) return [node.id, ...chain];
-		}
-
-		return null;
-	}
-
-	const chain = find(node);
-
-	if (chain) {
-		throw new Error(format_illegal_import_chain(chain));
-	}
-}
